+++ tests/TestSquirrelLifecycle.h
@@ -8,8 +8,14 @@ class TestSquirrelLifecycle : public QObject
     Q_OBJECT
 
 private slots:
+    void cleanup();
     void classification();
-    void updatePath();
+    void firstRunConsumption();
+    void layoutValidation();
+    void registryOwnershipDecisions();
+    void processResultContract();
+    void shortHelperEvidence();
+    void handleUsesExactOwnedSeams();
 };
 
 #endif // KEEPASSXC_TESTSQUIRRELLIFECYCLE_H